refactor(compensation): organize files by journey (onboarding/)
Move all onboarding-specific Compensation files (Compensation, EditCompensation,
JobsList, CompensationFlowComponents, compensationStateMachine, and the
co-located test) into a new onboarding/ subdirectory. The shared/ directory
(useJobForm, useCompensationForm) stays in place so partner hook exports are
unaffected.

This mirrors the layout established by PaymentMethod (onboarding/ + management/
+ shared/) and clears the way for a separate management/ tree to land in
subsequent PRs without churning the onboarding code.

No behavior change. Compensation/index.ts now re-exports through
onboarding/index.ts, so all external consumers (Employee/index.ts,
OnboardingFlow, src/index.ts hook exports) continue to resolve unchanged.
