+++ src/components/Employee/Compensation/index.ts
@@ -1,1 +1,1 @@
-export * from './Compensation'
+export * from './onboarding'
