+++ src/components/Employee/Compensation/onboarding/index.ts
@@ -1,0 +1,1 @@
+export * from './Compensation'
